+++ actix-server/src/handle.rs
@@ -46,6 +46,7 @@ impl ServerHandle {
         let _ = self.cmd_tx.send(ServerCommand::Stop {
             graceful,
             completion: Some(tx),
+            force_exit: false,
         });
 
         async {